% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-to-df.R
\name{arrays_to_df}
\alias{arrays_to_df}
\title{Combine arrays into data frame}
\usage{
arrays_to_df(x, dimnames_merge = c("y", "a"))
}
\arguments{
\item{x}{list of arrays to combine}

\item{dimnames_merge}{Character, vector of names of dimensions to merge by}
}
\value{
Merged data frame
}
\description{
Combine arrays into data frame
}
\examples{
listarr <- list(
  "H_star" = ex_H_star,
  "tau_dot_M" = ex_tau_dot_M,
  "phi_dot_M" = ex_phi_dot_M,
  "r" = ex_r,
  "phi_dot_E" = ex_phi_dot_E,
  "Q" = ex_Q
    )
 ERA_w <- arrays_to_df(listarr)




}
