% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-df-list.R
\name{combine_df_list}
\alias{combine_df_list}
\title{Combine data frames of results}
\usage{
combine_df_list(x, includes_ages = TRUE)
}
\arguments{
\item{x}{List of data frames to merge.}

\item{includes_ages}{Logical, do data frames include an age column to merge by?}
}
\value{
Data frame of merged results.
}
\description{
Combine data results from run reconstruction, merges by return year, population, and (if present) age.
}
\examples{
S_star1 <- get_S_star(E_star = ex_E_star, B_star = ex_B_star)
W_star1 <- get_W_star( S_star = S_star1$S_star, H_star = ex_H_star)
df_list <- list( S_star1$df, W_star1$df)
combine_df_list( x = df_list )


}
