% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nbinomial_params.R
\name{estimate_nbinomial_params}
\alias{estimate_nbinomial_params}
\title{Estimate parameters for binomial sampler}
\usage{
estimate_nbinomial_params(X)
}
\arguments{
\item{X}{Numeric, assumed to be comprised of all observed count data (e.g. K)
used for estimating mu, size, and prob parameters for a negative binomial sampler.}
}
\value{
List with three elements: (1) Numeric, mu parameter. (2) Numeric, size parameter. (3) Numeric, prob, mu / (size / mu).
}
\description{
Uses maximum likelihood estimation to find the most likely parameters to
use to populate a negative binomial sampler.
This was designed to accomodate the K parameter, but can be generalized
to sample from negative binomial probability mass functions.
The results of this function can be directly supplied to \code{rnbinomial()} when
performing Monte carlo sampling.
}
\examples{
params <- estimate_nbinomial_params(ex_k$kitsumkalum_escapement)
}
