% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-A-P.R
\docType{data}
\name{ex_A_P}
\alias{ex_A_P}
\title{Example ocean pre-fishery abundance for testing and examples}
\format{
\subsection{\code{ex_A_P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_A_P()
}
\usage{
ex_A_P
}
\description{
Example ocean pre-fishery abundance, which is the preterminal post-fishery
abundance plus the fish caught in preterminal non-net fisheries (e.g., troll).
}
\keyword{datasets}
