% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-B.R
\docType{data}
\name{ex_B}
\alias{ex_B}
\title{Example brood stock removal data by year for testing and examples}
\format{
\subsection{\code{ex_B}}{

A vector with one dimension:
\describe{
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_B
}
\description{
Example brood stock removal data (fake) for testing and examples.
}
\keyword{datasets}
