% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-B-star.R
\docType{data}
\name{ex_B_star}
\alias{ex_B_star}
\title{Example brood stock removal data by year and age for testing and examples}
\format{
\subsection{\code{ex_B_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_B_star
}
\description{
Example brood stock removal data (fake) for testing and examples.
}
\keyword{datasets}
