% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-E.R
\docType{data}
\name{ex_E}
\alias{ex_E}
\title{Example escapement by population and year for testing and examples}
\format{
\subsection{\code{ex_E}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_E()
}
\usage{
ex_E
}
\description{
Escapement of Chinook by population and year.
}
\keyword{datasets}
