% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-E-star.R
\docType{data}
\name{ex_E_star}
\alias{ex_E_star}
\title{Example escapement by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_E_star}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_E_star()
}
\usage{
ex_E_star
}
\description{
Escapement of Chinook by population, year, and age.
}
\keyword{datasets}
