% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-G.R
\docType{data}
\name{ex_G}
\alias{ex_G}
\title{Example gill-net catch data for testing and examples}
\format{
\subsection{\code{ex_G}}{

An array with two dimensions:
\describe{
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, make_P_G()
}
\usage{
ex_G
}
\description{
Example gill-net test fishery data by week and year.
}
\keyword{datasets}
