% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-H.R
\docType{data}
\name{ex_H}
\alias{ex_H}
\title{Example hatchery origin spawner data by year for testing and examples}
\format{
\subsection{\code{ex_H}}{

A vector with one dimension:
\describe{
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_H
}
\description{
Example hatchery origin spawner data (fake) for testing and examples.
}
\keyword{datasets}
