% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-H-star.R
\docType{data}
\name{ex_H_star}
\alias{ex_H_star}
\title{Example hatchery origin spawner data by year and age for testing and examples}
\format{
\subsection{\code{ex_H_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_H_star
}
\description{
Example hatchery origin spawner data (fake) for testing and examples.
}
\keyword{datasets}
