% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-k.R
\docType{data}
\name{ex_k}
\alias{ex_k}
\title{Example Kitsumkalum escapement data by year for testing and examples}
\format{
\subsection{\code{ex_k}}{

A dataframe with three variables:
\describe{
\item{year}{Year}
\item{kitsumkalum_escapement}{Escapement of Kitsumkalum Chinook ages 4-7, no jacks, estimated using mark-recapture and POPAN.}
\item{sd}{Standard Deviation of kitsumkalum_escapement.}
}
}
}
\source{
data-raw/make-example-data.R, data-raw/kitsumkalum-escapement.csv
}
\usage{
ex_k
}
\description{
Example Kitsumkalum escapement data by year
}
\keyword{datasets}
