% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-K-star.R
\docType{data}
\name{ex_K_star}
\alias{ex_K_star}
\title{Example Kitsumkalum escapement data by year and age for testing and examples}
\format{
\subsection{\code{ex_K_star}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, data-raw/K_star.csv
}
\usage{
ex_K_star
}
\description{
Example Kitsumkalum escapement data by year and age
}
\keyword{datasets}
