% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-MatureRun.R
\docType{data}
\name{ex_MatureRun}
\alias{ex_MatureRun}
\title{Example mature run by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_MatureRun}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_MatureRun()
}
\usage{
ex_MatureRun
}
\description{
Escapement of Chinook mature run data
}
\keyword{datasets}
