% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-n.R
\docType{data}
\name{ex_n}
\alias{ex_n}
\title{Example data of number of fish of population, year, and age for testing and examples}
\format{
\subsection{\code{ex_n}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_n
}
\description{
Values are the number of fish of age a assigned to population i caught in return year y. Includes ages 4-7.
}
\keyword{datasets}
