% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-N-total-run.R
\docType{data}
\name{ex_N_total_run}
\alias{ex_N_total_run}
\title{Example data of total run by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_N_total_run}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_N_total_run
}
\description{
Values are the total run in population i caught in return year y. Includes ages 4-7.
}
\keyword{datasets}
