% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-omega.R
\docType{data}
\name{ex_omega}
\alias{ex_omega}
\title{Example age proportion data by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_omega}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_omega()
}
\usage{
ex_omega
}
\description{
Values are age proportions by population i caught in return year y of age a. Includes ages 4-7.
}
\keyword{datasets}
