% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-omega-KM.R
\docType{data}
\name{ex_omega_KM}
\alias{ex_omega_KM}
\title{Example age proportion data for Kitsumkalum males by year, and age for testing and examples}
\format{
\subsection{\code{ex_omega_KM}}{

An array with three dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_omega_KM
}
\description{
Values are age proportions return year y of age a. Includes ages 4-7.
}
\keyword{datasets}
