% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-P.R
\docType{data}
\name{ex_P}
\alias{ex_P}
\title{Example genetic proportion data for testing and examples}
\format{
\subsection{\code{ex_P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, make_P_G()
}
\usage{
ex_P
}
\description{
Example genetic proportion data (fake) for testing and examples.
}
\keyword{datasets}
