% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-P-tilde.R
\docType{data}
\name{ex_P_tilde}
\alias{ex_P_tilde}
\title{Example pooled genetic proportion data for testing and examples}
\format{
\subsection{\code{ex_P_tilde}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_P_tilde()
}
\usage{
ex_P_tilde
}
\description{
Example genetic proportion data (fake) for testing and examples. An array of pooled genetic proportions with 2 dimensions: i (population) and y (year).
}
\keyword{datasets}
