% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-p-wild.R
\docType{data}
\name{ex_p_wild}
\alias{ex_p_wild}
\title{Example data of proportion wild spawners (natural origin)}
\format{
\subsection{\code{ex_p_wild}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_p()
}
\usage{
ex_p_wild
}
\description{
Array of wild spawner proportions with values between 0 and 1, with three dimensions: population (i), year (y), and age (a).
}
\keyword{datasets}
