% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-phi-dot-E.R
\docType{data}
\name{ex_phi_dot_E}
\alias{ex_phi_dot_E}
\title{Example data of preterminal total mortality exploitation rates}
\format{
\subsection{\code{ex_phi_dot_E}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_phi_dot_E
}
\description{
Array of preterminal total mortality exploitation rates
for non-net fisheries (e.g., troll), with two dimensions: year (y) and age (a).
From the CTC Chinook model output for Kitsumkalum CWT releases (KLM).
Equal to the number of fish caught plus incidental mortalities for non-net fisheries (e.g., troll), in
areas seaward of the terminal area, divided by total stock for the cohort, in nominal fish (not adjusted for AEQs).
}
\keyword{datasets}
