% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-phi-dot-M.R
\docType{data}
\name{ex_phi_dot_M}
\alias{ex_phi_dot_M}
\title{Example data of preterminal total mortality harvest rates rates of mature fish from net fisheries.}
\format{
\subsection{\code{ex_phi_dot_M}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_phi_dot_M
}
\description{
Array of preterminal net total mortality harvest rates on mature fish from net fisheries (greater than or equal to 0, less than 1) with two dimensions: year (y) and age (a).
Preterminal net total mortality harvest rate on mature fish: seine and gillnet catch and incidental mortalities of mature fish. From CTC Chinook model output.
}
\keyword{datasets}
