% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-phi-N.R
\docType{data}
\name{ex_phi_N}
\alias{ex_phi_N}
\title{Example data of preterminal fishing mortality in nominal fish}
\format{
\subsection{\code{ex_phi_N}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_phi_N()
}
\usage{
ex_phi_N
}
\description{
Array of preterminal fishing mortality in nominal fish, with three dimensions:
population (i), year (y), and age (a).
}
\keyword{datasets}
