% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-phi-Q.R
\docType{data}
\name{ex_phi_Q}
\alias{ex_phi_Q}
\title{Example data of preterminal fishing mortality in adult equivalents}
\format{
\subsection{\code{ex_phi_Q}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_phi_Q()
}
\usage{
ex_phi_Q
}
\description{
Array of preterminal fishing mortality in
adult equivalents, with three dimensions: population (i), year (y), and age (a).
}
\keyword{datasets}
