% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-Q.R
\docType{data}
\name{ex_Q}
\alias{ex_Q}
\title{Example data of adult equivalency rates}
\format{
\subsection{\code{ex_Q}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_Q
}
\description{
Array of adult equivalency (AEQ) rates with two dimensions: year (y) and age (a).
Output from CTC Chinook model based on Kitsumkalum CWT releases (KLM).
}
\keyword{datasets}
