% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-r.R
\docType{data}
\name{ex_r}
\alias{ex_r}
\title{Example data of maturation rates}
\format{
\subsection{\code{ex_r}}{

An array with two dimensions:
\describe{
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_r
}
\description{
Array of maturation rate by year and age. CTC Chinook model output for Kitsumkalum (KLM).
}
\keyword{datasets}
