% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-S.R
\docType{data}
\name{ex_S}
\alias{ex_S}
\title{Example data of spawners}
\format{
\subsection{\code{ex_S}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_S()
}
\usage{
ex_S
}
\description{
Array of spawner values for Chinook with two dimensions: population (i), and year (y).
}
\keyword{datasets}
