% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-S-star.R
\docType{data}
\name{ex_S_star}
\alias{ex_S_star}
\title{Example spawners by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_S_star}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_S_star()
}
\usage{
ex_S_star
}
\description{
Chinook spawners by population, year, and age.
}
\keyword{datasets}
