% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-sigma-P.R
\docType{data}
\name{ex_sigma_P}
\alias{ex_sigma_P}
\title{Example data of standard error for genetic proportion data for testing and examples}
\format{
\subsection{\code{ex_sigma_P}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{w}{Week}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, make_P_G()
}
\usage{
ex_sigma_P
}
\description{
Example standard error of genetic proportion data (fake) by population, week, and year.
}
\keyword{datasets}
