% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-sigma-P-tilde.R
\docType{data}
\name{ex_sigma_P_tilde}
\alias{ex_sigma_P_tilde}
\title{Example data of the standard error of pooled genetic proportions for testing and examples}
\format{
\subsection{\code{ex_sigma_P_tilde}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_P_tilde()
}
\usage{
ex_sigma_P_tilde
}
\description{
Example standard error of genetic proportions for testing and examples. An array of standard errors of pooled genetic proportions with 2 dimensions: i (population) and y (year).
}
\keyword{datasets}
