% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-Tau.R
\docType{data}
\name{ex_Tau}
\alias{ex_Tau}
\title{Example data for freshwater terminal mortalities in the Skeena River (lower and upper)}
\format{
\subsection{\code{ex_Tau}}{

An array with three dimensions:
\describe{
\item{y}{Year}
\item{tyee}{Numeric, vector of the total catch of adult and jack Chinook at the Skeena Tyee Test fishery, by year.}
\item{rec_catch_L}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
\item{rec_release_L}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
\item{FN_catch_L}{Numeric, vector of First Nations FSC catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
\item{rec_catch_U}{Numeric, vector of recreational catch of jack and adult Chinook in the upper Skeena (upstream of Terrace), by year.}
\item{FN_catch_U}{Numeric, vector of First Nations FSC catch of jack and adult Chinook in the upper Skeena (upstream of Terrace), by year.}
}
}
}
\source{
data-raw/make-example-data.R
}
\usage{
ex_Tau
}
\description{
Dataframe with example data for test fishery, recreational, and First Nations FSC catch in the Skeena River.
}
\keyword{datasets}
