% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-tau-L.R
\docType{data}
\name{ex_tau_L}
\alias{ex_tau_L}
\title{Example data for freshwater terminal moratlity in the lower Skeena}
\format{
\subsection{\code{ex_tau_L}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_tau_L()
}
\usage{
ex_tau_L
}
\description{
Total terminal mortalities of Chinook salmon in the lower Skeena River
(downstream of Terrace), by population, year, and age.
Array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).
}
\keyword{datasets}
