% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-Tau-L-total.R
\docType{data}
\name{ex_Tau_L_total}
\alias{ex_Tau_L_total}
\title{Example data of total terminal mortality in the lower Skeena}
\format{
\subsection{\code{ex_Tau_L_total}}{

A vector with one dimension:
\describe{
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_Tau_L_total()
}
\usage{
ex_Tau_L_total
}
\description{
A vector of total terminal mortalities in the lower Skeena (downstream of Terrace), by year.
Combination of Tyee test fishery, recreational, and First Nations FSC.
Note uppercase Tau denotes total not broken out by population or age.
Also note that jacks have been removed from this data.
}
\keyword{datasets}
