% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-tau-M.R
\docType{data}
\name{ex_tau_M}
\alias{ex_tau_M}
\title{Example data for marine terminal mortality}
\format{
\subsection{\code{ex_tau_M}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_tau_M()
}
\usage{
ex_tau_M
}
\description{
Array of total terminal mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).
Total terminal mortality (catch plus incidental mortality) in marine areas.
These include recreational and commercial catch near the Skeena River mouth, including Area 4.
}
\keyword{datasets}
