% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-tau-obsolete.R
\docType{data}
\name{ex_tau_obsolete}
\alias{ex_tau_obsolete}
\title{Example data for terminal mortality, obsolete from Winther et al. 2024}
\format{
\subsection{\code{ex_tau_obsolete}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_tau_obsolete()
}
\usage{
ex_tau_obsolete
}
\description{
Array of terminal mortalities with three dimensions:  population (i), year (y), and age (a).
}
\keyword{datasets}
