% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-tau-U.R
\docType{data}
\name{ex_tau_U}
\alias{ex_tau_U}
\title{Example data for freshwater terminal moratlity in the upper Skeena}
\format{
\subsection{\code{ex_tau_U}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_tau_U()
}
\usage{
ex_tau_U
}
\description{
Total terminal mortalities of Chinook salmon in the upper Skeena River
(upstream of Terrace), by population, year, and age.
Array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).
}
\keyword{datasets}
