% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-Tau-U-total.R
\docType{data}
\name{ex_Tau_U_total}
\alias{ex_Tau_U_total}
\title{Example data of total terminal mortality in the upper Skeena}
\format{
\subsection{\code{ex_Tau_U_total}}{

A vector with one dimension:
\describe{
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_Tau_U_total()
}
\usage{
ex_Tau_U_total
}
\description{
A vector of total terminal mortalities in the upper Skeena (upstream of Terrace), by year.
Combination of recrational and First Nations FSC.
Note uppercase Tau denotes total not broken out by population or age.
Also note that jacks have been removed from this data.
}
\keyword{datasets}
