% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-tau-W.R
\docType{data}
\name{ex_tau_W}
\alias{ex_tau_W}
\title{Example data for wild terminal mortality (freshwater and marine)}
\format{
\subsection{\code{ex_tau_W}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_tau_W()
}
\usage{
ex_tau_W
}
\description{
Array of wild terminal mortalities with three dimensions:  population (i), year (y), and age (a).
}
\keyword{datasets}
