% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-TermRun.R
\docType{data}
\name{ex_TermRun}
\alias{ex_TermRun}
\title{Example terminal run by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_TermRun}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_TermRun()
}
\usage{
ex_TermRun
}
\description{
Terminal run of Chinook
}
\keyword{datasets}
