% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-W-star.R
\docType{data}
\name{ex_W_star}
\alias{ex_W_star}
\title{Example wild spawners by population, year, and age for testing and examples}
\format{
\subsection{\code{ex_W_star}}{

An array with three dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
\item{a}{Age}
}
}
}
\source{
data-raw/make-example-data.R, get_W_star()
}
\usage{
ex_W_star
}
\description{
Chinook wild spawners by population, year, and age.
}
\keyword{datasets}
