% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-X.R
\docType{data}
\name{ex_X}
\alias{ex_X}
\title{Example data of Chinook return to Terrace}
\format{
\subsection{\code{ex_X}}{

An array with two dimensions:
\describe{
\item{i}{Population}
\item{y}{Year}
}
}
}
\source{
data-raw/make-example-data.R, get_X()
}
\usage{
ex_X
}
\description{
Array of Chinook return to Terrace values with two dimensions: population (i), and year (y).
}
\keyword{datasets}
