% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-A-P.R
\name{get_A_P}
\alias{get_A_P}
\title{Get ocean pre-fishery abundance}
\usage{
get_A_P(A_phi, phi_dot_E)
}
\arguments{
\item{A_phi}{Numeric, array of preterminal post-fishery abundance with
three dimensions: population (i), year (y), and age (a).}

\item{phi_dot_E}{Numeric, array of preterminal total mortality exploitation rates
for non-net fisheries (e.g., troll), with two dimensions: year (y) and age (a).
From the CTC Chinook model output for Kitsumkalum CWT releases (KLM).
Equal to the number of fish caught plus incidental mortalities for non-net fisheries (e.g., troll), in
areas seaward of the terminal area, divided by total stock for the cohort, in nominal fish (not adjusted for AEQs).}
}
\value{
List with two elements. First element: numeric, array of ocean pre-fishery
abundance with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element for plotting and tables.
}
\description{
Estimate ocean pre-fishery abundance, which is the preterminal post-fishery
abundance plus the fish caught in preterminal non-net fisheries (e.g., troll).
}
\examples{
  A_P <- get_A_P( A_phi = ex_A_phi, phi_dot_E = ex_phi_dot_E)


}
