% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-correct-age.R
\name{get_correct_age}
\alias{get_correct_age}
\title{Get total age from scale and Coded Wire Tag (CWT) ages}
\usage{
get_correct_age(scale_age, cwt_age)
}
\arguments{
\item{scale_age}{Numeric or integer, age from scales. Gilbert-Rich age (52 = total age of 5, etc.)}

\item{cwt_age}{Numeric or integer, age from Coded Wire Tag (CWT), equal to recovery year - brood year}
}
\value{
Integer, total age
}
\description{
When there are CWT ages and scale ages, uses CWT age if they are different.
}
\examples{

scale_age <- c(4,5,6,NA, 4, NA)
cwt_age <- c( NA, 4, 5, 5, NA, NA)
a <- get_correct_age( scale_age = scale_age, cwt_age = cwt_age )

}
