% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-E.R
\name{get_E}
\alias{get_E}
\title{Get escapement by population}
\usage{
get_E(
  K,
  X,
  Tau_U,
  known_population = "Kitsumkalum",
  aggregate_population = "Skeena",
  lower_populations = c("Lower Skeena", "Zymoetz-Fiddler"),
  upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"),
  save_csv = FALSE
)
}
\arguments{
\item{K}{Numeric, vector of estimates of Kitsumkalum Chinook spawners (ages 4, 5, 6, and 7, no jacks) based on a mark-recapture study and open population mark-recapture model (POPAN). One dimension: y (year). For more on POPAN models, see Cooch & White 2024 Chapter 12: http://www.phidot.org/software/mark/docs/book/pdf/chap12.pdf}

\item{X}{Numeric, array of returns to Terrace with two dimensions: population (i) and year (y).}

\item{Tau_U}{Integer, Chinook terminal mortalities from freshwater fisheries upsteam of Terrace, by year. Includes catch and incidental mortality estimate.}

\item{known_population}{Character, name of population i with known or estimated escapement. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena.}

\item{lower_populations}{Character vector, names of populations i downstream of Terrace. Defaults to Lower Skeena and Zymoetz-Fiddler.}

\item{upper_populations}{Character vector, names of populations i upstream of Terrace. Defaults to Upper Skeena, Middle Skeena, and Large Lakes.}

\item{save_csv}{If TRUE, save a csv of the data frame output.}
}
\value{
List with two elements. First element: numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculates E, escapement by population by subtracting freshwater terminal mortality from returns to Terrace if appropriate.
}
\examples{
E <- get_E(K = ex_k$kitsumkalum_escapement, X = ex_X$X, Tau_U = ex_Tau_U,
   known_population = "Kitsumkalum",
    aggregate_population = "Skeena",
    lower_populations = c("Lower Skeena", "Zymoetz-Fiddler"),
    upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"))

}
