% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-E-star.R
\name{get_E_star}
\alias{get_E_star}
\title{Get age-specific escapement}
\usage{
get_E_star(
  E,
  omega,
  use_alternate_escapement_by_age = TRUE,
  K_star,
  population_use_age_from_river_samples = "Kitsumkalum",
  add_6_7 = TRUE,
  save_csv = FALSE,
  save_location,
  save_name = "E_star.csv"
)
}
\arguments{
\item{E}{Numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y).}

\item{omega}{Numeric, array of proportions of each age with three dimensions: population (i), year (y), and age (a).}

\item{use_alternate_escapement_by_age}{Logical, should alternate escapement by age data be used for a population? Defaults to TRUE.}

\item{K_star}{Numeric, array of Kitsumkalum River escapement with two dimensions: return year (y) and age (a). Calculated using
age proportion data from spawning grounds (not from Tyee samples) for males and females separately.}

\item{population_use_age_from_river_samples}{Character, name of population to use alternate escapement by age data from sampling on spawning grounds. Defaults to Kitsumkalum.}

\item{add_6_7}{If TRUE, add age 7 fish to age 6 fish of the same return year (treat age 7 returns as age 6 returns from the age 6 return year). Default is TRUE. Note that this "modifies" the brood year of age 7 fish (true brood year +1).}

\item{save_csv}{If TRUE, save a csv of the data frame output.}

\item{save_location}{Sub-directory folder of where to save csv}

\item{save_name}{File name of csv to save. Defaults to "E_star.csv"}
}
\value{
List with two elements. First element: numeric, array of escapement values with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate escapement of Chinook by population, year, and age.
}
\examples{
  E_star <- get_E_star(E = ex_E, omega = ex_omega, K_star = ex_K_star, add_6_7 = TRUE)

}
