% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-K-star.R
\name{get_K_star}
\alias{get_K_star}
\title{Get age-specific escapement for Kitsumkalum escapement}
\usage{
get_K_star(
  K,
  y_K,
  K_M,
  K_F,
  omega_KM,
  omega_KF,
  H,
  H_star,
  save_csv = FALSE,
  save_location,
  save_name = "K_star.csv"
)
}
\arguments{
\item{K}{Numeric, escapement of total Kitsumkalum Chinook (males and females combined).}

\item{y_K}{Integer, vector of years for Kitsumkalum Chinook spawners K and sigma_K.}

\item{K_M}{Numeric, escapement of male Kitsumkalum Chinook.}

\item{K_F}{Numeric, escapement of female Kitsumkalum Chinook.}

\item{omega_KM}{Numeric, array of proportions of each age for males, with two dimensions: year (y), and age (a).}

\item{omega_KF}{Numeric, array of proportions of each age for females, with two dimensions: year (y), and age (a).}

\item{H}{numeric, number of hatchery fish in escapement by return year. From CTC model outputs.}

\item{H_star}{numeric, array of hatchery fish in escapement with two dimensions: year (y) and age (a). From CTC model outputs.}

\item{save_csv}{If TRUE, save a csv of the data frame output.}

\item{save_location}{Sub-directory folder of where to save csv}

\item{save_name}{File name of csv to save. Defaults to "E_star.csv"}
}
\value{
List with two elements. First element: numeric, array of Kitsumkalum River escapement with two dimensions: return year (y) and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate escapement of Chinook to the Kitsumkalum River by year and age. Uses age
proportions for males and females from scales collected during tagging program on the Kitsumkalum River (not Tyee test fishery).
Applies male and female age proportions to male and female escapement estimates from mark-recapture program and POPAN analysis.
Also uses age-specific hatchery component. Assumes sex ratio is equal in wild and hatchery fish in a given year.
}
\examples{
  K_star <- get_K_star(K = ex_k$kitsumkalum_escapement, y_K = ex_k$year,
                      K_M  = ex_k$male_escapement, K_F = ex_k$female_escapement,
                      omega_KM = ex_omega_KM, omega_KF = ex_omega_KF,
                      H = ex_H, H_star = ex_H_star )

}
