% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-p.R
\name{get_p}
\alias{get_p}
\title{Get proportion wild spawners}
\usage{
get_p(
  W_star,
  E_star,
  B_star,
  hatchery_population = "Kitsumkalum",
  aggregate_population = "Skeena"
)
}
\arguments{
\item{W_star}{Numeric, array of wild spawner values with three dimensions:  population (i), year (y), and age (a).}

\item{E_star}{Numeric, array of escapement values with three dimensions: population (i), year (y), and age (a).}

\item{B_star}{Numeric, array of brood stock removals with three dimensions: population (i), year (y), and age (a).}

\item{hatchery_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}
}
\value{
A list with two elements. First element: numeric, an array of wild spawner proportions with values between 0 and 1, with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculate the proportion of total escapement that are wild (natural origin).
}
\examples{
 p <- get_p(W_star = ex_W_star, E_star = ex_E_star, B_star = ex_B_star)

}
