% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-P-tilde.R
\name{get_P_tilde}
\alias{get_P_tilde}
\title{Get population proportions}
\usage{
get_P_tilde(P, G)
}
\arguments{
\item{P}{Numeric, array of genetic proportions with 3 dimensions: w (statistical week), i (population), and y (year)}

\item{G}{Numeric, array of how many Chinook were caught in the gillnet Tyee Test Fishery by week, with 2 dimensions: w (week) and y (year)}
}
\value{
Numeric, array of pooled genetic proportions with 2 dimensions: i (population) and y (year)
}
\description{
Do weekly expansions of catch by weekly genetic proportions, and then sum the weekly expansions by population and year
}
\examples{
library(rrandvec)
library(abind)
P <- make_P_G()[[1]]
G <- make_P_G()[[2]]

P_tilde <- get_P_tilde(P = P, G= G)

}
