% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-phi-N.R
\name{get_phi_N}
\alias{get_phi_N}
\title{Get preterminal fishing mortality in nominal fish}
\usage{
get_phi_N(A_P, A_phi)
}
\arguments{
\item{A_P}{Numeric, array of ocean pre-fishery
abundance with three dimensions: population (i), year (y), and age (a).}

\item{A_phi}{Numeric, array of preterminal post-fishery abundance with
three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of preterminal fishing mortality in nominal fish, with three dimensions:
population (i), year (y), and age (a). Second element: numeric, dataframe version of first element
}
\description{
The number of fish caught in pre-terminal non-net fisheries (e.g., troll).
}
\examples{
  phi_N <- get_phi_N(A_P = ex_A_P, A_phi = ex_A_phi)




}
