% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-R.R
\name{get_R}
\alias{get_R}
\title{Get recruits brood year}
\usage{
get_R(N)
}
\arguments{
\item{N}{Numeric, array of total run by population (i), year (y), and age (a).}
}
\value{
List with three elements. First element: numeric, array of recruits by  brood year with three dimensions: population (i), brood year (b).
Second element: data frame version of first element, plus return year. Third element: data frame of recruits by age, population, brood year and return year.
}
\description{
Calculate recruits by return year and brood year by adding up total run by brood year.
Produces an array with recruits by brood year, a data frame with both return year and brood year, and a data frame with ages and return year and brood year.
}
\details{
Get recruits by brood year

Calculate recruits by brood year, summing across ages.
}
\examples{
  R <- get_R( N = ex_N_total_run)

}
