% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-S.R
\name{get_S}
\alias{get_S}
\title{Get spawners by population}
\usage{
get_S(E, B, brood_population = "Kitsumkalum", aggregate_population = "Skeena")
}
\arguments{
\item{E}{numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y). Chinook that 'escaped' fisheries en route to spawning river.}

\item{B}{Numeric, vector of number of brood stock adult Chinook taken from the Kitsumkalum river, by year (y).}

\item{brood_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}
}
\value{
List with two elements. First element: numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculates S, spawners by population by subtracting brood stock removals from Kitsumkalum and Skeena aggregate populations. All other populations, spawners = escapement.
}
\examples{
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
 E <- array(data = sample(2000:3000, size = n_populations* n_years, replace=TRUE), dim = c(n_populations, n_years),
                       dimnames = list( i =  populations, y = years))
 E["Skeena", ] <- apply(E[!dimnames(E)$i=="Skeena", ], 2, sum)
 B <- sample(50:100, size=n_years)
 S <- get_S( E = E, B = B)

}
