% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-S-star.R
\name{get_S_star}
\alias{get_S_star}
\title{Get spawners by population, year, and age}
\usage{
get_S_star(
  E_star,
  B_star,
  aggregate_population = "Skeena",
  brood_population = "Kitsumkalum",
  save_csv = FALSE,
  save_location,
  save_name = "S_star.csv"
)
}
\arguments{
\item{E_star}{Numeric, array of escapement values with three dimensions: population (i), year (y), and age (a).}

\item{B_star}{Numeric, array of number of brood stock adult Chinook taken from the Kitsumkalum river, by year (y) and age (a).}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}

\item{brood_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{save_csv}{If TRUE, save a csv of the data frame output.}

\item{save_location}{Sub-directory folder of where to save csv}

\item{save_name}{File name of csv to save. Defaults to "S_star.csv"}
}
\value{
List with two elements. First element: numeric, array of spawner values with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element in long format, for plotting and tables.
}
\description{
From escapement, calculate the number of spawners by population, year, and age by subtracting brood stock removals for appropriate populations.
}
\examples{
  S_star <- get_S_star(E_star = ex_E_star, B_star = ex_B_star)


}
