% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stage.R
\name{get_stage}
\alias{get_stage}
\title{Get stage (adult or jack)}
\usage{
get_stage(
  scale_age,
  cwt_age,
  POH,
  comments,
  jack_ages = c("32", "31", "1M"),
  adult_ages = c("41", "42", "43", "51", "52", "53", "61", "62", "63", "71", "72", "73",
    "81", "82", "83", "2M", "3M", "4M", "5M"),
  jack_cutoff_length_POH = 450
)
}
\arguments{
\item{scale_age}{Character, vector of Gilbert-Rich scale ages (including partial ages) from Sclerochronology Lab.}

\item{cwt_age}{Numeric, vector of total ages from Coded Wire Tags (or a combination of scale ages and CWT ages). Gilbert-Rich total age (return year minus brood year).}

\item{POH}{Numeric, post-orbital hypural length, in mm}

\item{comments}{Character, vector of comments}

\item{jack_ages}{Character, vector of which scale_age values to include as jacks. Defaults to c("32", "31", "1M")}

\item{adult_ages}{Character, vector of which scale_age values to include as adults.
Defaults to c("41", "42", "43", "51", "52", "53", "61", "62", "63", "71", "72", "73",
"81", "82", "83", "2M", "3M", "4M", "5M" )}

\item{jack_cutoff_length_POH}{Numeric, single value of what POH length to use as a cutoff for jacks. Defaults to 450 mm.}
}
\value{
Character, "jack" or "adult"
}
\description{
Uses Skeena River Tyee Test Fishery biodata to assign stage of return (adult or jack) to Chinook, based on
CWT age, scale age, POH, and comments.
}
\details{
In skrunchy, the main purpose of this function is to filter the complete biodata to estimate
weekly catch of adult Chinook, which are used to do the weekly expansions of GSI proportions.
Default POH cutoff length is 450 mm. Note that this is what is used at Kitsumkalum.
This is more accurate when dividing between 32 and 42 male Chinook at Tyee, compared to 650 mm nose-fork length,
which categorizes a decent proportion of 42 males as jacks (incorrectly).
}
\examples{
scale_age <- c("31", "31", "32", "42", "52", "62", "1M", "2M", "3M", "4M", NA, NA)
cwt_age <- c( 4, 3, 3, 4, 5, 6, 3, 4, 5, 6, NA, NA)
POH <- c( 500, 440, 420, 500, 600, 700, 380, 500, 600, 700, NA, NA)
comments <- c( "NA", "Jack", "Jack", NA, NA, NA, NA, NA, NA, NA, "jack, bitten", "Jack")
d <- data.frame(scale_age, cwt_age, POH, comments)
d$revised_stage <- get_stage( scale_age = scale_age, cwt_age = cwt_age, POH = POH, comments = comments)


}
