% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stat-week.R
\name{get_stat_week}
\alias{get_stat_week}
\title{Get Alaska style statistical week from a date}
\usage{
get_stat_week(date)
}
\arguments{
\item{date}{Date, yyyy-mm-dd format date}
}
\value{
Integer, which stat week the date corresponds to.
}
\description{
This is used to group catch at Tyee into stat weeks, for weekly gill net catch
and as a utility function. See https://mtalab.adfg.alaska.gov/CWT/reports/sbp_calendar.aspx
}
\details{
This code is based on the "mixture_split.R" code, from Eric Rondeau and based on code by Colin ____
}
\examples{
d <- data.frame( "date" = seq.Date(ISOdate(2021,1,1), ISOdate(2025,12,31)))
d$stat_week <- get_stat_week(d$date)

}
