% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-T.R
\name{get_T}
\alias{get_T}
\title{Calculate Chinook returns to Terrace}
\usage{
get_T(P_tilde, sigma_P_tilde, K, sigma_K, y)
}
\arguments{
\item{P_tilde}{Numeric, array of genetic proportions at Skeena Tyee test fishery with 2 dimensions: i (population) and year (y).}

\item{sigma_P_tilde}{Numeric, array of SD of pooled genetic proportions with 2 dimensions: i (population) and y (year).}

\item{K}{Numeric, vector of estimates of Kitsumkalum Chinook spawners (ages 4, 5, 6, and 7, no jacks) based on a mark-recapture study and open population mark-recapture model (POPAN). One dimension: y (year). For more on POPAN models, see Cooch & White 2024 Chapter 12: http://www.phidot.org/software/mark/docs/book/pdf/chap12.pdf}

\item{sigma_K}{Numeric, vector of standard error (SE) of K. One dimension: y (year).}

\item{y}{Integer, vector of years for Kitsumkalum Chinook spawners K and sigma_K.}
}
\value{
Numeric, an array of
}
\description{
Calculate returns to Terrace of adult Chinook salmon for population i and year y. Uses pooled genetic proportions and the number of Chinook returns to the Kitsumkalum River.
}
\examples{



#' @export
}
