% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau.R
\name{get_tau}
\alias{get_tau}
\title{Get total terminal mortalities}
\usage{
get_tau(tau_U, tau_L, tau_M)
}
\arguments{
\item{tau_U}{Numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_L}{Numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_M}{Numeric, array of total terminal mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Add terminal mortalities in the upper Skeena, lower Skeena, and terminal marine areas.
}
\examples{
populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
n_populations <- length(populations)
years <- 2000:2001
n_years <- length(years)
ages <- c(3,4,5,6,7)
n_ages <- length(ages)
n_obs <- n_populations * n_years * n_ages
com_dims <- c(n_populations, n_years, n_ages)
com_dimnames <- list( i = populations, y= years, a = ages)
tau_U <- array( sample(0:1000, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames )
tau_L <- array( sample(0:500, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames)
tau_M <- array( sample(0:2000, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames)
tau <- get_tau( tau_U = tau_U, tau_L =tau_L, tau_M = tau_M)

}
