% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-Tau-L-total.R
\name{get_Tau_L_total}
\alias{get_Tau_L_total}
\title{Estimate Chinook total terminal mortalities in the lower Skeena River, by year.}
\usage{
get_Tau_L_total(
  omega_J,
  tyee,
  use_tyee,
  IM_tyee = 0.046,
  rec_catch_L,
  IM_rec_catch = 0.069,
  rec_release_L,
  IM_rec_release = 0.05,
  FN_catch_L,
  IM_FN_catch = 0.046,
  adult_ages = as.character(c(4, 5, 6, 7)),
  add_uncertainty = FALSE,
  cv_freshwater_mortality = 0.3
)
}
\arguments{
\item{omega_J}{Numeric, array of proportions of each age including jacks (age 3) for the Skeena River aggregate, with two dimensions: year (y), and age (a).
Note that this should include age proportion for Skeena aggregate. This is used to remove jacks from the terminal mortality.}

\item{tyee}{Numeric, vector of the total catch of adult and jack Chinook at the Skeena Tyee Test fishery, by year.}

\item{use_tyee}{Logical, should tyee test fishery mortality count data be used? Added since TNBC TERM N includes Tyee test fishery mortality.}

\item{IM_tyee}{Numeric, value to use for incidental mortality (drouput) for Tyee test fishery. Default is 4.6\%.}

\item{rec_catch_L}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}

\item{IM_rec_catch}{Numeric, value to use for incidental mortality (drop-off) for recreational catch. Default is 6.9\%.}

\item{rec_release_L}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}

\item{IM_rec_release}{Numeric, value to use for incidental mortality (release mortality) for fish caught and released in recreational fishery. Default is 5\%.}

\item{FN_catch_L}{Numeric, vector of First Nations FSC catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}

\item{IM_FN_catch}{Numeric, value to use for incidental mortality (drouput) for First Nations FSC catch. Default is 4.6\%.}

\item{adult_ages}{Character, vector of adult ages to use. Should match age names of omega_J columns. Defaults to "4", "5", "6", and "7".}

\item{add_uncertainty}{Logical, should a sample be drawn for adding uncertainty? Used for rru(). Defaults to FALSE}

\item{cv_freshwater_mortality}{Numeric, if add_uncertainty = TRUE, cv to apply to monte carlo sample draw of each source of mortality. Defaults to 0.3 (CV = 30\%).}
}
\value{
A numeric vector of total terminal mortalities including catch and incidental mortalities, for the lower Skeena (downstream of Terrace), by year.
}
\description{
Calculates the total terminal mortalities of Chinook salmon in the lower Skeena River
(downstream of Terrace), by year. Includes fishing plus incidental mortality.
Uses genetic data from Skeena Tyee test fishery, not CTC model outputs.
}
\details{
See Winther et al. 2024 Appendix 2 https://publications.gc.ca/site/eng/9.901355/publication.html and Cox-Rogers et al. 1999 Appendix 2, Table 1 https://publications.gc.ca/site/eng/9.805331/publication.html for more info on incidental mortality rates.
}
\examples{
omega_J_all <- get_omega( ex_n_with_jacks ) # get age proportions with jacks
omega_J <- omega_J_all$omega["Skeena",,] # select just Skeena age proportions
Tau_L_total <- get_Tau_L_total( omega_J = omega_J, tyee = ex_Tau$tyee,
                                use_tyee = FALSE,
                               rec_catch_L = ex_Tau$rec_catch_L,
                               rec_release_L = ex_Tau$rec_release_L,
                               FN_catch_L = ex_Tau$FN_catch_L )

}
