% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-M.R
\name{get_tau_M}
\alias{get_tau_M}
\title{Get total terminal mortality in marine areas}
\usage{
get_tau_M(W_star, tau_dot_M)
}
\arguments{
\item{W_star}{Numeric, array of wild spawner values for Chinook with two dimensions: population (i), year (y), and age (a).}

\item{tau_dot_M}{Numeric, array of terminal marine exploitation rates with dimensions year (y) and age (a). Sum of exploitation rate for terminal
marine net fisheries and exploitation rate for terminal marine sport fisheries
(coded "TNBC TERM N" and "TNBC TERM S" by the Chinook Technical Committee, respectively), which are outputs from the Kitsumkalum CWT analysis.}
}
\value{
List with two elements. First element: numeric, array of terminal total mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element for plotting and tables.
}
\description{
Estimate the total terminal mortality (catch plus incidental mortality) in marine areas.
These include recreational and commercial catch near the Skeena River mouth, including Area 4.
This is based on Coded Wire Tag data and Chinook Technical Committee model output.
}
\examples{
library(abind)
d <- make_P_G(start_year = 2000, end_year = 2001)
res <- get_P_tilde(P = d$P, sigma_P = d$sigma_P, G = d$G)
k <- data.frame( year = c(2000, 2001),
                kitsumkalum_escapement = c(5000, 6000),
                sd = c(500, 400))
X <- get_X(P_tilde = res$P_tilde, sigma_P_tilde = res$sigma_P_tilde, K= k$kitsumkalum_escapement,
          sigma_K = k$sd, y = k$year)
Tau_U <- sample(50:100, size = length(k$year), replace=TRUE)
E <- get_E(K = k$kitsumkalum_escapement, X = X$X, Tau_U = Tau_U,
   known_population = "Kitsumkalum",
    aggregate_population = "Skeena",
    lower_populations = c("Lower Skeena", "Zymoetz-Fiddler"),
    upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"))
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6,7)
  p_ages <- c(30,40,40,1)
  n_ages <- length(ages)
  # Make up some age data
  d <- sapply(p_ages, FUN = function(x){ rpois( n = n_populations*n_years, lambda= x) })
  n <- array( d,  dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  omega <- get_omega(n)
  K_star <- array(runif(n = n_years * n_ages, 5000, 20000), dim = c(n_years, n_ages), dimnames= list(y = years, a = ages))
  E_star <- get_E_star(E = E$E, omega = omega$omega, K_star = K_star)
  B_star <- array(sample(1:50, size= n_years*n_ages), dim = c(n_years, n_ages), dimnames = list( y = years, a = ages))
  S_star <- get_S_star(E_star = E_star$E_star, B_star = B_star)
  H_star <- array(sample(10:40, size=n_years*n_ages), dim = c(n_years, n_ages), dimnames = list( y = years, a = ages))
  W_star <- get_W_star(S_star = S_star$S_star, H_star = H_star,
    aggregate_population = "Skeena", hatchery_population = "Kitsumkalum")
 tau_dot_M <- array( data = sample(1:3, size=n_years*n_ages, replace =TRUE)/10, dim = c(n_years, n_ages),dimnames = list("y" = years, "a" = ages))
 tau_M <- get_tau_M( W_star = W_star$W_star, tau_dot_M = tau_dot_M )


}
