% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-obsolete.R
\name{get_tau_obsolete}
\alias{get_tau_obsolete}
\title{Get total terminal mortalities - obsolete function}
\usage{
get_tau_obsolete(tau_U, tau_L, tau_M)
}
\arguments{
\item{tau_U}{Numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_L}{Numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_M}{Numeric, array of total terminal mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of total terminal mortalities, with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Add terminal mortalities in the upper Skeena, lower Skeena, and terminal marine areas.
}
\examples{
tau <- get_tau_obsolete( tau_U = ex_tau_U, tau_L = ex_tau_L, tau_M = ex_tau_M)

}
