% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-W.R
\name{get_tau_W}
\alias{get_tau_W}
\title{Get wild terminal mortalities}
\usage{
get_tau_W(tau_U, tau_L, tau_M, p)
}
\arguments{
\item{tau_U}{Numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_L}{Numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_M}{Numeric, array of total terminal mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).}

\item{p}{Numeric, an array of wild spawner proportions with values between 0 and 1, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of wild terminal mortalities with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate the wild total terminal mortalities by population, year, and age.
}
\examples{
tau_W <- get_tau_W(tau_L = ex_tau_L, tau_U = ex_tau_U, tau_M = ex_tau_M, p = ex_p_wild)


}
