% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-W-obsolete.R
\name{get_tau_W_obsolete}
\alias{get_tau_W_obsolete}
\title{Get wild terminal mortalities (obsolete version of function)}
\usage{
get_tau_W_obsolete(tau, p)
}
\arguments{
\item{tau}{Numeric, array of total terminal mortalities, with three dimensions: population (i), year (y), and age (a).}

\item{p}{Numeric, an array of wild spawner proportions with values between 0 and 1, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of wild terminal mortalities with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate the wild total terminal mortalities by population, year, and age.
}
\examples{
tau_W <- get_tau_W_obsolete(tau = ex_tau, p = ex_p_wild)


}
