% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-W-reproduce.R
\name{get_tau_W_reproduce}
\alias{get_tau_W_reproduce}
\title{Get wild terminal mortalities (different formula to reproduce old excel analysis)}
\usage{
get_tau_W_reproduce(tau_U, tau_L, tau_M, p)
}
\arguments{
\item{tau_U}{Numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_L}{Numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).}

\item{tau_M}{Numeric, array of total terminal mortalities in marine fisheries, with three dimensions: population (i), year (y), and age (a).}

\item{p}{Numeric, an array of wild spawner proportions with values between 0 and 1, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of wild terminal mortalities with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate the wild total terminal mortalities by population, year, and age.
}
\details{
Winther et al. 2024 applied p to τU and τL, whereas in the updated run reconstruction, we
apply p to τL only. This is because τU is assumed to be all wild spawnwers,
since all hatchery origin spawners in this method are Kitsumkalum origin and
we assume exit the Skeena mainstem just downstream of Terrace, and thus
are not encountered in freshwater fisheries upstream of Terrace. Therefore,
we assume that all fish caught in freshwater fisheries upstream of Terrace
(τU) are wild origin.
}
\examples{
tau_W <- get_tau_W_reproduce(tau_L = ex_tau_L, tau_U = ex_tau_U, tau_M = ex_tau_M, p = ex_p_wild)


}
