% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-total-age.R
\name{get_total_age}
\alias{get_total_age}
\title{Get total Gilbert-Rich age}
\usage{
get_total_age(GR_age)
}
\arguments{
\item{GR_age}{Character, vector of Gilbert-Rich ages as they come from the SCL (sclerochronology lab).}
}
\value{
Integer, vector of total age. For example. a 52 fish becomes 5.
}
\description{
Uses the Gilbert-Rich age from the Scale age lab, and returns the total age (integer).
Partial ages from regrowth scales (e.g., 3M, 2M) become NAs.
}
\examples{

raw_ages <- c("32", "42", "52", "51", "62", "2M", "RG", "1M")
age <- get_total_age(raw_ages)

}
