% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-TotalRun.R
\name{get_TotalRun}
\alias{get_TotalRun}
\title{Get TotalRun by return year}
\usage{
get_TotalRun(MatureRun, phi_Q)
}
\arguments{
\item{MatureRun}{Numeric, array of mature run size by population (i), year (y), and age (a).}

\item{phi_Q}{Numeric, array of preterminal fishing mortality in
adult equivalents, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with three elements. First element: numeric, array of recruits by return year with three dimensions: population (i), return year (y), and
age (a). Third element: data frame version of first and second elements, for plotting and tables.
}
\description{
Calculate the total run by return year by adding mature run and preterminal fishing mortality
in adult equivalents.
This is all the fish in the cohort: harvest of all types plus escapement. Used to estimate recruitment.
Produces an array by return year and a data frame with both return year and brood year.
}
\examples{
  TotalRun <- get_TotalRun( MatureRun = ex_MatureRun, phi_Q = ex_phi_Q)

}
