% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-W-star.R
\name{get_W_star}
\alias{get_W_star}
\title{Get wild spawners by age}
\usage{
get_W_star(
  S_star,
  H_star,
  aggregate_population = "Skeena",
  hatchery_population = "Kitsumkalum"
)
}
\arguments{
\item{S_star}{Numeric, array of spawner values with three dimensions:  population (i), year (y), and age (a).}

\item{H_star}{Numeric, array of number of hatchery origin spawners returning to the population with a hatchery (e.g., Kitsumkalum River), by year (y) and age (a). Excludes age 3 fish (jacks).}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}

\item{hatchery_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}
}
\value{
List with two elements. First element: numeric, array of wild spawner values with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element in long format, for plotting and tables.
}
\description{
Calculates wild spawners, specific by age at return. Uses total wild spawners by year and population, and age proportions from Tyee.
}
\examples{
  W_star <- get_W_star(S_star = ex_S_star, H_star = ex_H_star,
    aggregate_population = "Skeena", hatchery_population = "Kitsumkalum")

}
