% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-X.R
\name{get_X}
\alias{get_X}
\title{Calculate Chinook returns to Terrace}
\usage{
get_X(
  P_tilde,
  sigma_P_tilde,
  K,
  y_K,
  sigma_K,
  known_population = "Kitsumkalum",
  aggregate_population = "Skeena",
  save_csv = FALSE
)
}
\arguments{
\item{P_tilde}{Numeric, array of genetic proportions at Skeena Tyee test fishery with 2 dimensions: i (population) and year (y).}

\item{sigma_P_tilde}{Numeric, array of SD of pooled genetic proportions with 2 dimensions: i (population) and y (year).}

\item{K}{Numeric, vector of estimates of Kitsumkalum Chinook spawners (ages 4, 5, 6, and 7, no jacks) based on a mark-recapture study and open population mark-recapture model (POPAN). One dimension: y (year). For more on POPAN models, see Cooch & White 2024 Chapter 12: http://www.phidot.org/software/mark/docs/book/pdf/chap12.pdf}

\item{y_K}{Integer, vector of years for Kitsumkalum Chinook spawners K and sigma_K.}

\item{sigma_K}{Numeric, vector of standard error (SE) of K. One dimension: y (year).}

\item{known_population}{Name of population i with known or estimated escapement. Defaults to Kitsumkalum.}

\item{aggregate_population}{Name of population i that is the sum of the other populations. Defaults to Skeena.}

\item{save_csv}{If TRUE, save a csv of the data frame output.}
}
\value{
A list with three elements. First element: Numeric, X which is an array of returns to Terrace with two dimensions: population (i) and year (y).
Second element: numeric, sigma_X which is an array of SE of returns to Terrace of Chinook, with two dimensions: population (i) and year (y).
Third element: dataframe with X, sigma_X, and year merged for plotting and reporting.
}
\description{
Calculate returns to Terrace of adult Chinook salmon (X) for population i and year y. Uses pooled genetic proportions and the number of Chinook returns to the Kitsumkalum River.
}
\examples{
X <- get_X(P_tilde = ex_P_tilde, sigma_P_tilde = ex_sigma_P_tilde, K= ex_k$kitsumkalum_escapement,
            y_K = ex_k$year, sigma_K = ex_k$sd)

}
