% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-P-G.R
\name{make_P_G}
\alias{make_P_G}
\title{Make fake genetic data and catch data}
\usage{
make_P_G(
  catch_range = c(0, 100),
  n_weeks = 12,
  n_years = 40,
  start_year = 1980,
  population_names = c("Kitsumkalum", "Lower Skeena", "Middle Skeena", "Zymoetz-Fiddler",
    "Large Lakes", "Upper Skeena")
)
}
\arguments{
\item{catch_range}{Integer, length 2 vector of minimum and maximum catch values to sample between.}

\item{n_weeks}{Number of weeks fished.}

\item{n_years}{Number of years of data to create.}

\item{start_year}{Start year for fake data.}

\item{population_names}{Character vector with names of populatins.}
}
\value{
List with two elements. The first element is P, an array of genetic proportions with 3 dimensions: w (week), i (population), and y (year). The second element is G, an array of how many Chinook were caught in the gillnet Tyee Test Fishery by week, with 2 dimensions: w (week) and y (year).
}
\description{
Make two fake data arrays, one for genetic mixture analysis by week, and one weekly test fishery catch.
}
\examples{
library(rrandvec)
library(abind)
res <- make_P_G()
P <- res[[1]]
G <- res[[2]]

}
