% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-rates.R
\name{process_rates}
\alias{process_rates}
\title{Process rate data by replacing high values with a rolling mean value.}
\usage{
process_rates(rate_array, upper_limit = 0.5, roll_mean_window = 5)
}
\arguments{
\item{rate_array}{Numeric, array of exploitation rates. Two dimensions: y (return year) and a (age).}

\item{upper_limit}{Numeric, upper limit used to determine which rates are dropped and replaced with rolling mean. Default is 0.5.}

\item{roll_mean_window}{Numeric, how many values to use when applying rolling mean. Default is 5.}
}
\value{
Array, high values replaced with means.
}
\description{
Takes rate data from CTC exploitation rate analysis (or any rate data) and
applies a threshold of a high value to it. Values above this upper limit are
replaced with a rolling mean that is calculated after dropping high values.
}
\examples{
years <- 1984:2020
ages <- 4:6
d <- matrix( rbeta( n = length(years)*length(ages), shape1 = 2, shape2 = 5 ), nrow = length(years))
names(d)
dimnames(d) <- list( "y" = years, "a" = ages )
dn <- process_rates( d )
plot( d ~ dn )

}
