% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_exploitation_rates.R
\name{resample_exploitation_rates}
\alias{resample_exploitation_rates}
\title{Resample exploitation rates with added uncertainty, for Monte Carlo run reconstruction}
\usage{
resample_exploitation_rates(
  data,
  rate_resample_method = c("lognormal", "beta", "logit"),
  rate_resample_cv = 0.3,
  rate_resample_phi = 50,
  rate_resample_sigma = 0.4,
  rate_resample_max_er = 0.5
)
}
\arguments{
\item{data}{Numeric, array (matrix) of exploitation rates to sample. Two dimensions:
First dimension: return year y. Second dimension: return age a.}

\item{rate_resample_method}{Character, which method to use to add error. Options are "lognormal", "beta", or "logit".}

\item{rate_resample_cv}{Numeric, decimal CV to use for lognormal method.}

\item{rate_resample_phi}{Numeric, phi to use for beta method.}

\item{rate_resample_sigma}{Numeric, sigma to use for logit method.}

\item{rate_resample_max_er}{Numeric, single value of max ER rate from resampling. Defaults to 0.5}
}
\value{
Numeric, array (matrix) of new exploitation rates
}
\description{
Resample exploitation rates with added uncertainty, for Monte Carlo run reconstruction
}
\examples{
resample_exploitation_rates( data = ex_phi_dot_E,
 rate_resample_method = "lognormal",
 rate_resample_cv = 0.3,
 rate_resample_max_er = 0.5 )


}
