% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-name-key.R
\docType{data}
\name{variable_name_key}
\alias{variable_name_key}
\title{Variable names and details}
\format{
\subsection{\code{variable_name_key}}{

A dataframe with three variables:
\describe{
\item{skrunchy}{Names of variables in code}
\item{detail}{More detailed names}
\item{skrunchy_with_detail}{First two columns combined}
}
}
}
\source{
data-raw/misc.R, data-raw/variable_name_key.csv
}
\usage{
variable_name_key
}
\description{
Key of variable names and more details, used for producing results tables with more descriptive names
}
\keyword{datasets}
