% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-A-P.R
\name{get_A_P}
\alias{get_A_P}
\title{Get ocean pre-fishery abundance}
\usage{
get_A_P(A_phi, phi_dot_E)
}
\arguments{
\item{A_phi}{Numeric, array of preterminal post-fishery abundance with
three dimensions: population (i), year (y), and age (a).}

\item{phi_dot_E}{Numeric, array of preterminal total mortality exploitation rates
for non-net fisheries (e.g., troll), with two dimensions: year (y) and age (a).
From the CTC Chinook model output for Kitsumkalum CWT releases (KLM).
Equal to the number of fish caught plus incidental mortalities for non-net fisheries (e.g., troll), in
areas seaward of the terminal area, divided by total stock for the cohort, in nominal fish (not adjusted for AEQs).}
}
\value{
List with two elements. First element: numeric, array of ocean pre-fishery
abundance with three dimensions: population (i), year (y), and age (a).
Second element: data frame version of first element for plotting and tables.
}
\description{
Estimate ocean pre-fishery abundance, which is the preterminal post-fishery
abundance plus the fish caught in preterminal non-net fisheries (e.g., troll).
}
\examples{
 populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6,7)
  n_ages <- length(ages)
  A_phi<- array(sample(5000:10000, size=n_populations*n_years*n_ages),
    dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  A_phi["Skeena",, ] <- apply(A_phi[ !dimnames(A_phi)$i =="Skeena",, ], c(2,3), sum)
  phi_dot_E <- array( sample( seq(0.01, 0.2, 0.01), size = n_years * n_ages, replace=TRUE), dim = c(n_years, n_ages),
    dimnames = list( y = years, a = ages))
  A_P <- get_A_P( A_phi = A_phi, phi_dot_E = phi_dot_E)


}
