% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-A-phi.R
\name{get_A_phi}
\alias{get_A_phi}
\title{Get preterminal post-fishery abundance}
\usage{
get_A_phi(MatureRun, r)
}
\arguments{
\item{MatureRun}{Numeric, array of mature run size by population (i), year (y), and age (a).}

\item{r}{Numeric, array of maturation rate by year and age. CTC Chinook model output for Kitsumkalum (KLM).}
}
\value{
List with two elements. First element: numeric, array of preterminal post-fishery abundance with
three dimensions: population (i), year (y), and age (a). Second element: data frame version
of first element for plotting and tables.
}
\description{
Calculate the preterminal post-fishery abundance, which is the abundance of the
cohort after non-net fisheries (e.g., troll), after accounting for maturation rate but
before preterminal net (gillnet, seine) fishery harvest of mature fish. Of the fish that survived
non-net ocean fisheries, the number of fish that matured into the mature run.
}
\examples{
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6,7)
  n_ages <- length(ages)
  MatureRun <- array(sample(5000:10000, size=n_populations*n_years*n_ages),
    dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  MatureRun["Skeena",, ] <- apply(MatureRun[ !dimnames(MatureRun)$i =="Skeena",, ], c(2,3), sum)
  r <- array( sample( seq(0.8,0.9, 0.01), size = n_years * n_ages, replace=TRUE), dim = c(n_years, n_ages),
    dimnames = list( y = years, a = ages))
  r[, dimnames(r)$a \%in\% c(6,7) ] <- 1
  A_phi <- get_A_phi(MatureRun = MatureRun, r = r)


}
