% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-MatureRun.R
\name{get_MatureRun}
\alias{get_MatureRun}
\title{Get mature run size abundance}
\usage{
get_MatureRun(TermRun, phi_dot_M)
}
\arguments{
\item{TermRun}{Numeric, array of wild terminal run abundance, with dimensions population (i), year (y), and age (a).}

\item{phi_dot_M}{Numeric, array of preterminal net total mortality harvest rates (greater than or equal to 0, less than 1) with two dimensions: year (y) and age (a).
Preterminal net total mortality harvest rate on mature fish: seine and gillnet catch and incidental mortalities of mature fish.}
}
\value{
List with two elements. First element: numeric, array of mature run size by population (i), year (y), and age (a).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculates the mature run size abundance, by population, year, and age. The mature run is equal to the wild terminal run plus
the mature fish that were caught in marine net (gillnet and seine) fisheries plus incidental mortalities (preterminal). Net fisheries are considered to encounter mature
fish so adjustments for natural mortality and adult equivalent calculations are not necessary.
}
\examples{
library(abind)
d <- make_P_G(start_year = 2000, end_year = 2001)
res <- get_P_tilde(P = d$P, sigma_P = d$sigma_P, G = d$G)
k <- data.frame( year = c(2000, 2001),
                kitsumkalum_escapement = c(5000, 6000),
                sd = c(500, 400))
X <- get_X(P_tilde = res$P_tilde, sigma_P_tilde = res$sigma_P_tilde, K= k$kitsumkalum_escapement,
          sigma_K = k$sd, y = k$year)
Tau_U <- sample(50:100, size = length(k$year), replace=TRUE)
E <- get_E(K = k$kitsumkalum_escapement, X = X$X, Tau_U = Tau_U,
   known_population = "Kitsumkalum",
    aggregate_population = "Skeena",
    lower_populations = c("Lower Skeena", "Zymoetz-Fiddler"),
    upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"))
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6)
  p_ages <- c(30,40,40)
  n_ages <- length(ages)
  # Make up some age data
  d <- sapply(p_ages, FUN = function(x){ rpois( n = n_populations*n_years, lambda= x) })
  n <- array( d,  dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  omega <- get_omega(n)
  E_star <- get_E_star(E = E$E, omega = omega$omega)
  B_star <- array(sample(1:50, size= n_years*n_ages), dim = c(n_years, n_ages), dimnames = list( y = years, a = ages))
  #B_star[,4] <- 0 # make age 7 brood = 0 so you don't get negative fish in S_star
  S_star <- get_S_star(E_star = E_star$E_star, B_star = B_star)
  H_star <- array(sample(10:40, size=n_years*n_ages), dim = c(n_years, n_ages), dimnames = list( y = years, a = ages))
  W_star <- get_W_star(S_star = S_star$S_star, H_star = H_star,
    aggregate_population = "Skeena", hatchery_population = "Kitsumkalum")
n_obs <- n_populations * n_years * n_ages
com_dims <- c(n_populations, n_years, n_ages)
com_dimnames <- list( i = populations, y= years, a = ages)
tau_U <- array( sample(0:1000, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames )
tau_L <- array( sample(0:500, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames)
tau_M <- array( sample(0:2000, size = n_obs, replace=TRUE), dim = com_dims, dimnames = com_dimnames)
tau <- get_tau( tau_U = tau_U, tau_L =tau_L, tau_M = tau_M)
p <- array( runif(length(tau$tau), 0.8, 1), dim= dim(tau$tau), dimnames = dimnames(tau$tau))
tau_W <- get_tau_W(tau = tau$tau, p = p)
TermRun <- get_TermRun(tau_W = tau_W$tau_W, W_star = W_star$W_star, B_star=B_star)
use_arr <- TermRun$TermRun["Kitsumkalum",,]
phi_dot_M <- array(runif(length(use_arr), 0.01, 0.3), dim = dim(use_arr), dimnames = dimnames(use_arr))
MatureRun <- get_MatureRun(TermRun = TermRun$TermRun, phi_dot_M = phi_dot_M)




}
