% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-p.R
\name{get_p}
\alias{get_p}
\title{Get proportion wild spawners}
\usage{
get_p(
  W,
  E,
  hatchery_population = "Kitsumkalum",
  aggregate_population = "Skeena"
)
}
\arguments{
\item{W}{Numeric, matrix of wild spanwer values for Chinook with two dimensions: population (i), and year (y).}

\item{E}{Numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y). Chinook that 'escaped' fisheries en route to spawning river.}

\item{hatchery_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}
}
\value{
A list with two elements. First element: numeric, a matrix of wild spawner proportions with values between 0 and 1, with two dimensions: population (i) and year (y).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculate the proportion of total spawners that are wild spawners (natural origin).
}
\examples{
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
 E <- array(data = sample(2000:3000, size = n_populations* n_years, replace=TRUE), dim = c(n_populations, n_years),
                       dimnames = list( i =  populations, y = years))
 E["Skeena", ] <- apply(E[!dimnames(E)$i=="Skeena", ], 2, sum)
 B <- sample(50:100, size=n_years)
 S <- get_S( E = E, B = B)
 H <- sample(500:600, size=n_years)
 W <- get_W( S = S$S, H = H)
 p <- get_p( W = W$W, E = E)

}
