% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-phi-N.R
\name{get_phi_N}
\alias{get_phi_N}
\title{Get preterminal fishing mortality in nominal fish}
\usage{
get_phi_N(A_P, A_phi)
}
\arguments{
\item{A_P}{Numeric, array of ocean pre-fishery
abundance with three dimensions: population (i), year (y), and age (a).}

\item{A_phi}{Numeric, array of preterminal post-fishery abundance with
three dimensions: population (i), year (y), and age (a).}
}
\value{
List with two elements. First element: numeric, array of preterminal fishing mortality in nominal fish, with three dimensions:
population (i), year (y), and age (a). Second element: numeric, dataframe version of first element
}
\description{
The number of fish caught in pre-terminal non-net fisheries (e.g., troll).
}
\examples{
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6,7)
  n_ages <- length(ages)
  MatureRun <- array(sample(5000:10000, size=n_populations*n_years*n_ages),
    dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  MatureRun["Skeena",, ] <- apply(MatureRun[ !dimnames(MatureRun)$i =="Skeena",, ], c(2,3), sum)
  r <- array( sample( seq(0.8,0.9, 0.01), size = n_years * n_ages, replace=TRUE), dim = c(n_years, n_ages),
    dimnames = list( y = years, a = ages))
  r[, dimnames(r)$a \%in\% c(6,7) ] <- 1
  A_phi <- get_A_phi(MatureRun = MatureRun, r = r)
  phi_dot_E <- array( sample( seq(0.01, 0.2, 0.01), size = n_years * n_ages, replace=TRUE), dim = c(n_years, n_ages),
    dimnames = list( y = years, a = ages))
  A_P <- get_A_P( A_phi = A_phi$A_phi, phi_dot_E = phi_dot_E)
  phi_N <- get_phi_N(A_P = A_P$A_P, A_phi = A_phi$A_phi)




}
