% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-phi-Q.R
\name{get_phi_Q}
\alias{get_phi_Q}
\title{Get preterminal fishing mortality in adult equivalents}
\usage{
get_phi_Q(phi_N, Q)
}
\arguments{
\item{phi_N}{Numeric, array of preterminal fishing mortality in nominal fish, with three dimensions:
population (i), year (y), and age (a).}

\item{Q}{Numeric, array of adult equivalency (AEQ) rates with two dimensions: year (y) and age (a).
Output from CTC Chinook model based on Kitsumkalum CWT releases (KLM).}
}
\value{
List with two elements. First element: numeric, array of preterminal fishing mortality in
adult equivalents, with three dimensions: population (i), year (y), and age (a).
Second element: dataframe version of first element, for plotting and tables.
}
\description{
Calculate the preteriminal fishing mortality in adult equivalents (AEQ).
}
\examples{
  phi_Q <- get_phi_Q(phi_N = ex_phi_N, Q = ex_Q)


}
