% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-R.R
\name{get_R}
\alias{get_R}
\title{Get recruits by brood year}
\usage{
get_R(R_star_b, R_star_df)
}
\arguments{
\item{R_star_b}{Numeric, array of recruits by brood year, with three dimensions: population (i), brood year (b), and age (a).}

\item{R_star_df}{Numeric, data frame of recruits by poopulation, return year, brood year, and age.}
}
\value{
List with two elements. First element: numeric, array of recruits with two dimensions: population (i) and brood year (b).
Second element: data frame version of first element, for plotting and tables.
}
\description{
Calculate recruits by brood year, summing across ages.
}
\examples{
  R <- get_R(R_star_b = ex_R_star$R_star_b, R_star_df = ex_R_star$df)

}
