% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-R-star.R
\name{get_R_star}
\alias{get_R_star}
\title{Get recruits by return year and brood year}
\usage{
get_R_star(MatureRun, phi_Q)
}
\arguments{
\item{MatureRun}{Numeric, array of mature run size by population (i), year (y), and age (a).}

\item{phi_Q}{Numeric, array of preterminal fishing mortality in
adult equivalents, with three dimensions: population (i), year (y), and age (a).}
}
\value{
List with three elements. First element: numeric, array of recruits by return year with three dimensions: population (i), return year (y), and
age (a). Second element: numeric, array of recruits by brood year with three dimensions: population (i), brood year (b), and
age (a). Third element: data frame version of first and second elements, for plotting and tables.
}
\description{
Calculate recruits by return year and brood year by adding mature run and preterminal fishing mortality
in adult equivalents. Produces an array by return year, an array by brood year, and a data frame with both return year and brood year.
}
\examples{
  R_star <- get_R_star( MatureRun = ex_MatureRun, phi_Q = ex_phi_Q)

}
