% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-S.R
\name{get_S}
\alias{get_S}
\title{Get spawners by population}
\usage{
get_S(
  method = "sum_ages",
  S_star,
  E,
  B,
  brood_population = "Kitsumkalum",
  aggregate_population = "Skeena"
)
}
\arguments{
\item{method}{How to calcute spawners (S)? If "sum_ages", take the sum across ages of S_star (age-specific spawner array).
If "subtract_brood", uses non-age-specific escapement (E) and subtracts non-age-specific brood stock removals (B). Defaults to sum_ages.}

\item{S_star}{Numeric, array of spawner values with three dimensions:  population (i), year (y), and age (a).}

\item{E}{numeric, matrix of escapement values for Chinook with two dimensions: population (i), and year (y). Chinook that 'escaped' fisheries en route to spawning river.}

\item{B}{Numeric, vector of number of brood stock adult Chinook taken from the Kitsumkalum river, by year (y).}

\item{brood_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}
}
\value{
List with two elements. First element: numeric, matrix of spawner values for Chinook with two dimensions: population (i), and year (y).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculates spawners (S,  not age-specific) by population by either summing S_star by ages, or subtracting brood stock removals (B) from escapement (E) of Kitsumkalum and Skeena aggregate populations.
All other populations, spawners (S) = escapement (E).
}
\examples{
# Sum ages method
S <- get_S(method="sum_ages", S_star = ex_S_star)

# Subtract brood method
 S <- get_S( method= "subtract_brood", E = ex_E, B = ex_B)

}
