% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-S-star.R
\name{get_S_star}
\alias{get_S_star}
\title{Get spawners by population, year, and age}
\usage{
get_S_star(
  E_star,
  B_star,
  aggregate_population = "Skeena",
  brood_population = "Kitsumkalum"
)
}
\arguments{
\item{E_star}{Numeric, array of escapement values with three dimensions: population (i), year (y), and age (a).}

\item{B_star}{Numeric, array of number of brood stock adult Chinook taken from the Kitsumkalum river, by year (y) and age (a).}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}

\item{brood_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}
}
\value{
List with two elements. First element: numeric, array of spawner values with three dimensions:  population (i), year (y), and age (a).
Second element: data frame version of first element in long format, for plotting and tables.
}
\description{
From escapement, calculate the number of spawners by population, year, and age by subtracting brood stock removals for appropriate populations.
}
\examples{
library(abind)
d <- make_P_G(start_year = 2000, end_year = 2001)
res <- get_P_tilde(P = d$P, sigma_P = d$sigma_P, G = d$G)
k <- data.frame( year = c(2000, 2001),
                kitsumkalum_escapement = c(5000, 6000),
                sd = c(500, 400))
X <- get_X(P_tilde = res$P_tilde, sigma_P_tilde = res$sigma_P_tilde, K= k$kitsumkalum_escapement,
          sigma_K = k$sd, y = k$year)
Tau_U <- sample(50:100, size = length(k$year), replace=TRUE)
E <- get_E(K = k$kitsumkalum_escapement, X = X$X, Tau_U = Tau_U,
   known_population = "Kitsumkalum",
    aggregate_population = "Skeena",
    lower_populations = c("Lower Skeena", "Zymoetz-Fiddler"),
    upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"))
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
  ages <- c(4,5,6,7)
  p_ages <- c(30,40,40,1)
  n_ages <- length(ages)
  # Make up some age data
  d <- sapply(p_ages, FUN = function(x){ rpois( n = n_populations*n_years, lambda= x) })
  n <- array( d,  dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
  omega <- get_omega(n)
  E_star <- get_E_star(E = E$E, omega = omega$omega)
  B_star <- array(sample(1:50, size= n_years*n_ages), dim = c(n_years, n_ages), dimnames = list( y = years, a = ages))
  B_star[,4] <- 0 # make age 7 brood = 0 so you don't get negative fish in S_star
  S_star <- get_S_star(E_star = E_star$E_star, B_star = B_star)


}
