% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-L.R
\name{get_tau_L}
\alias{get_tau_L}
\title{Estimate Chinook total terminal mortalities in the lower Skeena River}
\usage{
get_tau_L(
  Tau_L,
  omega,
  P_tilde,
  aggregate_population = "Skeena",
  add_6_7 = TRUE
)
}
\arguments{
\item{Tau_L}{Numeric, vector of total terminal mortalities in the lower Skeena (downstream of Terrace), by year. Combination of FSC and recreational. Note uppercase Tau denotes total not broken out by population or age.}

\item{omega}{Numeric, array of proportions of each age with three dimensions: population (i), year (y), and age (a).}

\item{P_tilde}{Numeric, array of genetic proportions at Skeena Tyee test fishery with 2 dimensions: i (population) and year (y).}

\item{aggregate_population}{Character, name of aggregate population}

\item{add_6_7}{Logical, If TRUE, add age 7 fish to age 6 fish of the same brood year (treat age 7 mortalities as age 6 mortalities from the age 6 brood year). Default is TRUE. Note that this "modifies" the return year of age 7 mortalities (true return year -1).}
}
\value{
A list with two objects. First object: numeric, array of total terminal mortalities in the lower Skeena with three dimensions: population (i), year (y), and age (a).
Second object: data frame with same data in long format (columns for population, year, and age) for plotting.
}
\description{
Calculates the total terminal mortalities of Chinook salmon in the lower Skeena River
(downstream of Terrace), by population, year, and age. Includes fishing plus incidental mortality.
Uses genetic data from Skeena Tyee test fishery, not CTC model outputs.
}
\examples{
tau_L <- get_tau_L( Tau_L = ex_Tau_L_total, omega = ex_omega,
                P_tilde = ex_P_tilde, aggregate_population = "Skeena",
                add_6_7 = TRUE)

}
