% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tau-U.R
\name{get_tau_U}
\alias{get_tau_U}
\title{Estimate Chinook total terminal mortalities in the upper Skeena River}
\usage{
get_tau_U(
  Tau_U,
  omega,
  P_tilde,
  aggregate_population = "Skeena",
  upper_populations = c("Middle Skeena", "Large Lakes", "Upper Skeena"),
  lower_populations = c("Lower Skeena", "Kitsumkalum", "Zymoetz-Fiddler")
)
}
\arguments{
\item{Tau_U}{Numeric, vector of total terminal mortalities in the lupper Skeena (upstream of Terrace), by year. Combination of FSC and recreational. Note uppercase Tau denotes total not broken out by population or age.}

\item{omega}{Numeric, array of proportions of each age with three dimensions: population (i), year (y), and age (a).}

\item{P_tilde}{Numeric, array of genetic proportions at Skeena Tyee test fishery with 2 dimensions: i (population) and year (y).}

\item{aggregate_population}{Character, name of aggregate population. Defaults to "Skeena"}

\item{upper_populations}{Character, vector of names of upper populations. Defaults to Middle Skeena, Large Lakes, and Upper Skeena.}

\item{lower_populations}{Character, vector of names of lower populations. Defaults to Lower Skeena, Kitsumkalum, and Zymoetz-Fiddler.}
}
\value{
A list with two objects. First object: numeric, array of total terminal mortalities in the upper Skeena with three dimensions: population (i), year (y), and age (a).
Second object: data frame with same data in long format (columns for population, year, and age) for plotting.
}
\description{
Calculates the total terminal mortalities of Chinook salmon in the upper Skeena River
(upstream of Terrace), by population, year, and age. Includes fishing plus incidental mortality.
}
\examples{
populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
n_populations <- length(populations)
years <- 2000:2001
n_years <- length(years)
ages <- c(3,4,5,6,7)
p_ages <- c(20,30,40,40,1)
n_ages <- length(ages)
d <- sapply(p_ages, FUN = function(x){ rpois( n = n_populations*n_years, lambda= x) })
n <- array( d,  dim = c(n_populations, n_years, n_ages), dimnames = list(i = populations, y = years, a = ages))
omega <- get_omega(n)
Tau_U <- sample(500:1000, size=n_years)
P_G <-  make_P_G(, start_year = 2000, n_years = n_years, population_names = populations[1:6])
P_tilde <- get_P_tilde( P = P_G$P, sigma_P = P_G$sigma_P, G = P_G$G)
tau_U <- get_tau_U( Tau_U = Tau_U, omega = omega$omega, P_tilde = P_tilde$P_tilde,
    aggregate_population = "Skeena", upper_populations = c("Middle Skeena", "Large Lakes", "Upper Skeena"),
    lower_populations = c("Lower Skeena", "Kitsumkalum", "Zymoetz-Fiddler") )

}
