% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-W.R
\name{get_W}
\alias{get_W}
\title{Get wild spawners (natural origin) by population}
\usage{
get_W(
  S,
  H,
  hatchery_population = "Kitsumkalum",
  aggregate_population = "Skeena"
)
}
\arguments{
\item{S}{numeric, matrix of spawner values for Chinook with two dimensions: population (i), and year (y). Chinook that spawned (doesn't include brood stock removals).}

\item{H}{Numeric, vector of number of hatchery origin spawners returning to the population with a hatchery (e.g., Kitsumkalum River), by year (y).}

\item{hatchery_population}{Character, name of population i with where brood removals occurred. Defaults to Kitsumkalum.}

\item{aggregate_population}{Character, name of population i that is the sum of the other populations. Defaults to Skeena. Brood also need to removed from this since brood population is part of the aggregate.}
}
\value{
List with two elements. First element: numeric, array of wild spawner values for Chinook with two dimensions: population (i), and year (y).
Second element: data frame version of matrix, for plotting and tables.
}
\description{
Calculates W, wild spawners by population by subtracting hatchery-origin spawners from Kitsumkalum and Skeena aggregate populations. All other populations, wild spawners = spawners.
}
\examples{
  populations <- c("Kitsumkalum", "Lower Skeena", "Zymoetz-Fiddler", "Upper Skeena", "Middle Skeena", "Large Lakes", "Skeena")
  n_populations <- length(populations)
  years <- 2000:2001
  n_years <- length(years)
 E <- array(data = sample(2000:3000, size = n_populations* n_years, replace=TRUE), dim = c(n_populations, n_years),
                       dimnames = list( i =  populations, y = years))
 E["Skeena", ] <- apply(E[!dimnames(E)$i=="Skeena", ], 2, sum)
 B <- sample(50:100, size=n_years)
 S <- get_S( E = E, B = B)
 H <- sample(500:600, size=n_years)
 W <- get_W( S = S$S, H = H)
}
