% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-X.R
\name{get_X}
\alias{get_X}
\title{Calculate Chinook returns to Terrace}
\usage{
get_X(
  P_tilde,
  sigma_P_tilde,
  K,
  sigma_K,
  y,
  known_population = "Kitsumkalum",
  aggregate_population = "Skeena"
)
}
\arguments{
\item{P_tilde}{Numeric, array of genetic proportions at Skeena Tyee test fishery with 2 dimensions: i (population) and year (y).}

\item{sigma_P_tilde}{Numeric, array of SD of pooled genetic proportions with 2 dimensions: i (population) and y (year).}

\item{K}{Numeric, vector of estimates of Kitsumkalum Chinook spawners (ages 4, 5, 6, and 7, no jacks) based on a mark-recapture study and open population mark-recapture model (POPAN). One dimension: y (year). For more on POPAN models, see Cooch & White 2024 Chapter 12: http://www.phidot.org/software/mark/docs/book/pdf/chap12.pdf}

\item{sigma_K}{Numeric, vector of standard error (SE) of K. One dimension: y (year).}

\item{y}{Integer, vector of years for Kitsumkalum Chinook spawners K and sigma_K.}

\item{known_population}{Name if population i with known or estimated escapement. Defaults to Kitsumkalum.}

\item{aggregate_population}{Name of population i that is the sum of the other populations. Defaults to Skeena.}
}
\value{
A list with three elements. First element: Numeric, X which is an array of returns to Terrace with two dimensions: population (i) and year (y).
Second element: numeric, sigma_X which is an array of SE of returns to Terrace of Chinook, with two dimensions: population (i) and year (y).
Third element: dataframe with X, sigma_X, and year merged for plotting and reporting.
}
\description{
Calculate returns to Terrace of adult Chinook salmon for population i and year y. Uses pooled genetic proportions and the number of Chinook returns to the Kitsumkalum River.
}
\examples{
library(abind)
d <- make_P_G(start_year = 2000, end_year = 2001)
res <- get_P_tilde(P = d$P, sigma_P = d$sigma_P, G = d$G)
k <- data.frame( year = c(2000, 2001),
                kitsumkalum_escapement = c(5000, 6000),
                sd = c(500, 400))
X <- get_X(P_tilde = res$P_tilde, sigma_P_tilde = res$sigma_P_tilde, K= k$kitsumkalum_escapement,
          sigma_K = k$sd, y = k$year)
X

}
