% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-P-G.R
\name{make_P_G}
\alias{make_P_G}
\title{Make fake genetic data and catch data}
\usage{
make_P_G(
  catch_range = c(0, 100),
  n_weeks = 12,
  n_years = NA,
  start_year = 1984,
  end_year = 2023,
  population_names = c("Kitsumkalum", "Lower Skeena", "Middle Skeena", "Zymoetz-Fiddler",
    "Large Lakes", "Upper Skeena"),
  seed = TRUE,
  save_csv = TRUE
)
}
\arguments{
\item{catch_range}{Integer, length 2 vector of minimum and maximum catch values to sample between.}

\item{n_weeks}{Number of weeks fished.}

\item{n_years}{Number of years of data to create. Default is NA}

\item{start_year}{Start year for fake data.}

\item{end_year}{End year for fake data.}

\item{population_names}{Character vector with names of populations.}

\item{seed}{Set seed to 1?}
}
\value{
List with three elements. The first element is P, an array of genetic proportions with 3 dimensions: i (population), w (week), and y (year).
The second element is sigma_P, an array of SD of the genetic proportions, with 3 dimensions: i (population), w (week),  and y (year).
The third element is G, an array of how many Chinook were caught in the gillnet Tyee Test Fishery by week, with 2 dimensions: w (week) and y (year).
}
\description{
Make three fake data arrays: (1) proportions and (2) SD from genetic mixture analysis by population, week, and year,
and (3) test fishery Chinook catch by week and year.
}
\examples{
res <- make_P_G()
P <- res$P
sigma_P <- res$sigma_P
G <- res$G

}
