% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rru_iterate.R
\name{rru_iterate}
\alias{rru_iterate}
\title{Do run reconstruction with uncertainty, iterate many times}
\usage{
rru_iterate(n_iter = 1000, seed = 1, ...)
}
\arguments{
\item{n_iter}{Inter, number of iterations to run. Defaults to 1000.}

\item{seed}{Integer or NULL, should a seed be set? For reproducibiliy. Default to 1.}

\item{...}{
  Arguments passed on to \code{\link{rru}}
  \describe{
    \item{\code{n_gsi_samples}}{Integer, number of gsi samples by year, for sampling.}
    \item{\code{n_age_samples}}{Integer, number of age samples by year, for sampling.}
    \item{\code{n_age_samples_J}}{Integer, number of age samples by year, including age 3 (jacks), for sampling.}
    \item{\code{iteration_number}}{Integer, index of iteration}
    \item{\code{name_key}}{Dataframe, with details on names}
    \item{\code{save_outputs}}{Logical, whether to save the outputs. Default is TRUE.}
    \item{\code{P}}{Numeric, array of genetic proportions with 3 dimensions: i (population), w (statistical week),and y (year)}
    \item{\code{sigma_P}}{Numeric, array of SD of the genetic proportions, with 3 dimensions: i (population), w (week), and y (year).}
    \item{\code{G}}{Numeric, array of how many Chinook were caught in the gillnet Tyee Test Fishery by week, with 2 dimensions: w (week) and y (year)}
    \item{\code{K}}{Numeric, vector of estimates of Kitsumkalum Chinook spawners (ages 4, 5, 6, and 7, no jacks) based on a mark-recapture study and open population mark-recapture model (POPAN). One dimension: y (year). For more on POPAN models, see Cooch & White 2024 Chapter 12: http://www.phidot.org/software/mark/docs/book/pdf/chap12.pdf}
    \item{\code{sigma_K}}{Numeric, vector of standard error (SE) of K. One dimension: y (year).}
    \item{\code{y_K}}{Integer, vector of years for Kitsumkalum Chinook spawners K and sigma_K.}
    \item{\code{known_population}}{Name of population i with known or estimated escapement. Defaults to Kitsumkalum.}
    \item{\code{aggregate_population}}{Name of population i that is the sum of the other populations. Defaults to Skeena.}
    \item{\code{omega_J}}{Numeric, array of proportions of each age including jacks (age 3) for the Skeena River aggregate, with two dimensions: year (y), and age (a).
Note that this should include age proportion for Skeena aggregate. This is used to remove jacks from the terminal mortality.}
    \item{\code{tyee}}{Numeric, vector of the total catch of adult and jack Chinook at the Skeena Tyee Test fishery, by year.}
    \item{\code{rec_catch_L}}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
    \item{\code{rec_release_L}}{Numeric, vector of recreational catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
    \item{\code{FN_catch_L}}{Numeric, vector of First Nations FSC catch of jack and adult Chinook in the lower Skeena (downstream of Terrace), by year.}
    \item{\code{rec_catch_U}}{Numeric, vector of recreational catch of jack and adult Chinook in the upper Skeena (upstream of Terrace), by year.}
    \item{\code{FN_catch_U}}{Numeric, vector of First Nations FSC catch of jack and adult Chinook in the upper Skeena (upstream of Terrace), by year.}
    \item{\code{lower_populations}}{Character vector, names of populations i downstream of Terrace. Defaults to Lower Skeena and Zymoetz-Fiddler.}
    \item{\code{upper_populations}}{Character vector, names of populations i upstream of Terrace. Defaults to Upper Skeena, Middle Skeena, and Large Lakes.}
    \item{\code{omega}}{Numeric, array of proportions of each age with three dimensions: population (i), year (y), and age (a).}
    \item{\code{K_star}}{Numeric, array of Kitsumkalum River escapement with two dimensions: return year (y) and age (a). Calculated using
age proportion data from spawning grounds (not from Tyee samples) for males and females separately.}
    \item{\code{add_6_7}}{If TRUE, add age 7 fish to age 6 fish of the same return year (treat age 7 returns as age 6 returns from the age 6 return year). Default is TRUE. Note that this "modifies" the brood year of age 7 fish (true brood year +1).}
    \item{\code{B_star}}{Numeric, array of number of brood stock adult Chinook taken from the Kitsumkalum river, by year (y) and age (a).}
    \item{\code{H_star}}{Numeric, array of number of hatchery origin spawners returning to the population with a hatchery (e.g., Kitsumkalum River), by year (y) and age (a). Excludes age 3 fish (jacks).
This is the sum of CTC model output column Esc_exp by return year and age, for KLM (fry) and KLY (yearlings) combined.}
    \item{\code{tau_dot_M}}{Numeric, array of terminal marine exploitation rates with dimensions year (y) and age (a). Sum of exploitation rate for terminal
marine net fisheries and exploitation rate for terminal marine sport fisheries
(coded "TNBC TERM N" and "TNBC TERM S" by the Chinook Technical Committee, respectively), which are outputs from the Kitsumkalum CWT analysis.}
    \item{\code{phi_dot_M}}{Numeric, array of preterminal net total mortality harvest rates (greater than or equal to 0, less than 1) with two dimensions: year (y) and age (a).
Preterminal net total mortality harvest rate on mature fish: seine and gillnet catch and incidental mortalities of mature fish. From CTC Chinook model output.}
    \item{\code{r}}{Numeric, array of maturation rate by year and age. CTC Chinook model output for Kitsumkalum (KLM).}
    \item{\code{phi_dot_E}}{Numeric, array of preterminal total mortality exploitation rates
for non-net fisheries (e.g., troll), with two dimensions: year (y) and age (a).
From the CTC Chinook model output for Kitsumkalum CWT releases (KLM).
Equal to the number of fish caught plus incidental mortalities for non-net fisheries (e.g., troll), in
areas seaward of the terminal area, divided by total stock for the cohort, in nominal fish (not adjusted for AEQs).}
    \item{\code{Q}}{Numeric, array of adult equivalency (AEQ) rates with two dimensions: year (y) and age (a).
Output from CTC Chinook model based on Kitsumkalum CWT releases (KLM).}
    \item{\code{use_tyee}}{Logical, should tyee test fishery mortality count data be used? Added since TNBC TERM N includes Tyee test fishery mortality.}
    \item{\code{add_uncertainty}}{Logical, should a sample be drawn for adding uncertainty? Used for rru(). Defaults to FALSE}
    \item{\code{cv_freshwater_mortality}}{Numeric, if add_uncertainty = TRUE, cv to apply to monte carlo sample draw of each source of mortality. Defaults to 0.3 (CV = 30\%).}
  }}
}
\value{
List of results.
}
\description{
See \code{\link[=rru]{rru()}} for a single run.
}
\examples{
 n_gsi_samples <- apply(ex_n, 2, sum)
 n_age_samples <- apply(ex_n, c(1,2), sum)
 n_age_samples_J <- apply(ex_n_with_jacks, c(1,2), sum)
 results <- rru_iterate(
   n_iter = 10,
   P = ex_P,
   sigma_P = ex_sigma_P,
   G = ex_G,
   K = ex_k$kitsumkalum_escapement,
   n_gsi_samples = n_gsi_samples,
   n_age_samples = n_age_samples,
   n_age_samples_J = n_age_samples_J,
   sigma_K = ex_k$sd,
   y_K = ex_k$year,
   omega = ex_omega,
   omega_J = ex_omega_J,
   tyee = ex_Tau$tyee,
   rec_catch_L = ex_Tau$rec_catch_L,
   rec_release_L = ex_Tau$rec_release_L,
   FN_catch_L = ex_Tau$FN_catch_L,
   rec_catch_U = ex_Tau$rec_catch_U,
   FN_catch_U = ex_Tau$FN_catch_U,
   known_population = "Kitsumkalum",
   aggregate_population = "Skeena",
   lower_populations = c("Lower Skeena", "Zymoetz"),
   upper_populations = c("Upper Skeena", "Middle Skeena", "Large Lakes"),
   K_star = ex_K_star,
   add_6_7 = TRUE,
   B_star = ex_B_star,
   H_star = ex_H_star,
   tau_dot_M = ex_tau_dot_M,
   phi_dot_M = ex_phi_dot_M,
   r = ex_r,
   phi_dot_E = ex_phi_dot_E,
   Q = ex_Q,
   name_key = variable_name_key,
   save_outputs = FALSE,
   )

}
